refactor data preprocessing
